tweet_collector/requirements.txt
tweepy 
SQLAlchemy
psycopg2-binary